Add tensor camera, posenet and push up detection
Tensor camera was embedded with expo camera to turn video frames into tensor images. Posenet, a pretrained pose detection model, then uses these tensor images to produce an output containing positions of different points of the body. The x and y positioning of the shoulder, elbow, and wrist is used to calculate the angle to determine 1 repetition of push up.